pyproject.toml
[project]
name = "subsystem-router-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.1",
    "pyyaml>=6.0.3",
    "pydantic>=2.0.0",
    "torch>=2.9.1",
    "transformers>=4.57.5",
    "scikit-learn>=1.8.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'"  },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
